Use a double for damage related keys.

+++ src/main/java/org/spongepowered/common/data/provider/item/stack/ArmorItemStackData.java
@@ -57,7 +57,7 @@ public final class ArmorItemStackData {
                     .create(Keys.DAMAGE_ABSORPTION)
                         .get(h -> {
                             if (h.getItem() instanceof ArmorItem) {
-                                return ((ArmorItem) h.getItem()).getDamageReduceAmount();
+                                return (double) ((ArmorItem) h.getItem()).getDamageReduceAmount();
                             }
                             return null;
                         })
